Interactive agent

+++ platform/lib/registry.py
@@ -7,7 +7,7 @@ import sys
 sys.path.append('./lib/')
 
 
-###### 
+######
 import time
 import argparse
 import logging
@@ -35,9 +35,9 @@ from redis.commands.search.query import Query
 #######
 import numpy as np
 from sentence_transformers import SentenceTransformer
- 
 
-class Registry():
+
+class Registry:
     def __init__(self, name="REGISTRY", id=None, sid=None, cid=None, prefix=None, suffix=None, properties={}):
 
         self.name = name
@@ -63,7 +63,6 @@ class Registry():
             if self.suffix:
                 self.cid = self.cid + ":" + self.suffix
 
-
         self._initialize(properties=properties)
 
         self._start()
@@ -73,10 +72,9 @@ class Registry():
         self._initialize_properties()
         self._update_properties(properties=properties)
 
-
     def _initialize_properties(self):
         self.properties = {}
-        
+
         # db connectivity
         self.properties['db.host'] = 'localhost'
         self.properties['db.port'] = 6379
@@ -85,8 +83,6 @@ class Registry():
         self.properties['vector_dimensions'] = 4
         # embeddings model
         self.properties['embeddings_model'] = 'paraphrase-MiniLM-L6-v2'
-
-
 
     def _update_properties(self, properties=None):
         if properties is None:
@@ -114,7 +110,7 @@ class Registry():
                 return None
             else:
                 if single:
-                    return(value[0])
+                    return value[0]
                 else:
                     return value
         else:
@@ -122,7 +118,7 @@ class Registry():
 
     def _init_registry_namespace(self):
         # create registry-specific registry
-        self.connection.json().set(self._get_data_namespace(), '$', {'contents':{}}, nx=True)
+        self.connection.json().set(self._get_data_namespace(), '$', {'contents': {}}, nx=True)
 
     def _get_index_name(self):
         return self.cid
@@ -147,7 +143,7 @@ class Registry():
 
             # schema
             schema = self._build_index_schema()
-            
+
             # index definition
             definition = IndexDefinition(prefix=[doc_prefix], index_type=IndexType.HASH)
 
@@ -161,23 +157,25 @@ class Registry():
         vector_dimensions = self.properties['vector_dimensions']
 
         schema = (
-                # name
-                TextField("name", weight=2.0),
-                # type 
-                TextField("type"), 
-                # scope 
-                TextField("scope"), 
-                # description text
-                TextField("description"),
-                # description embedding
-                VectorField("vector",                  
-                    "FLAT", {                          
-                        "TYPE": "FLOAT32",             
-                        "DIM": vector_dimensions,      
-                        "DISTANCE_METRIC": "COSINE",   
-                    }
-                ),
-            )
+            # name
+            TextField("name", weight=2.0),
+            # type
+            TextField("type"),
+            # scope
+            TextField("scope"),
+            # description text
+            TextField("description"),
+            # description embedding
+            VectorField(
+                "vector",
+                "FLAT",
+                {
+                    "TYPE": "FLOAT32",
+                    "DIM": vector_dimensions,
+                    "DISTANCE_METRIC": "COSINE",
+                },
+            ),
+        )
         return schema
 
     def build_index(self):
@@ -216,19 +214,12 @@ class Registry():
                 r = contents[key]
                 self._set_index_record(r, recursive=recursive, pipe=pipe)
 
-           
     def _create_index_doc(self, name, type, scope, description, pipe=None):
-        
-        #TODO: Identify the best way to compute embedding vector, for now name + description 
+
+        # TODO: Identify the best way to compute embedding vector, for now name + description
         vector = self._compute_embedding_vector(name + " " + description)
 
-        doc = {
-            'name': name,
-            'type': type,
-            'scope': scope,
-            'description': description,
-            'vector': vector
-        }
+        doc = {'name': name, 'type': type, 'scope': scope, 'description': description, 'vector': vector}
 
         # define key
         doc_key = self.__doc_key(name, type, scope)
@@ -242,9 +233,9 @@ class Registry():
 
     def __doc_key(self, name, type, scope):
         index_name = self._get_index_name()
-        doc_prefix = self._get_doc_prefix() 
+        doc_prefix = self._get_doc_prefix()
 
-        if scope[len(scope)-1] == '/':
+        if scope[len(scope) - 1] == '/':
             scope = scope[:-1]
 
         return doc_prefix + ':' + type + ":" + scope + "/" + name
@@ -260,9 +251,8 @@ class Registry():
 
         for key in contents:
             r = contents[key]
-           
-            self._delete_index_record(r, pipe=pipe)
 
+            self._delete_index_record(r, pipe=pipe)
 
     def _delete_index_doc(self, name, type, scope, pipe=None):
 
@@ -274,19 +264,18 @@ class Registry():
             pipe.hdel(doc_key, 1)
         else:
             pipe = self.connection.pipeline()
-            for field in ["name","type","scope","description","vector"]:
+            for field in ["name", "type", "scope", "description", "vector"]:
                 pipe.hdel(doc_key, field)
             res = pipe.execute()
             print(res)
 
-
     def search_records(self, keywords, type=None, scope=None, approximate=False, hybrid=False, page=0, page_size=5, page_limit=10):
-        
+
         index_name = self._get_index_name()
         doc_prefix = self._get_doc_prefix()
 
         q = None
-        
+
         qs = ""
 
         if type:
@@ -296,43 +285,21 @@ class Registry():
 
         if hybrid:
             q = "( " + qs + " " + " $kw " + " )" + " => [KNN " + str((page_limit) * page_size) + " @vector $v as score]"
-            
-            query = (
-                Query(q)
-                .sort_by("score")
-                .return_fields("id", "name", "type", "scope", "score")
-                .paging(0, page_limit * page_size)
-                .dialect(2)
-            )
+
+            query = Query(q).sort_by("score").return_fields("id", "name", "type", "scope", "score").paging(0, page_limit * page_size).dialect(2)
 
         else:
             if approximate:
                 if qs == "":
                     qs = "*"
-                q =  "( " + qs + " )" + " => [KNN " + str((page_limit) * page_size) + " @vector $v as score]"
-                query = (
-                    Query(q)
-                    .sort_by("score")
-                    .return_fields("id", "name", "type", "scope", "score")
-                    .paging(0, page_limit * page_size)
-                    .dialect(2)
-                )
+                q = "( " + qs + " )" + " => [KNN " + str((page_limit) * page_size) + " @vector $v as score]"
+                query = Query(q).sort_by("score").return_fields("id", "name", "type", "scope", "score").paging(0, page_limit * page_size).dialect(2)
 
             else:
                 q = "( " + qs + " " + " $kw " + " )"
-                query = (
-                    Query(q)
-                    .return_fields("id", "name", "type", "scope")
-                    .paging(0, page_limit * page_size)
-                    .dialect(2)
-                )
+                query = Query(q).return_fields("id", "name", "type", "scope").paging(0, page_limit * page_size).dialect(2)
 
-        
-
-        query_params = {
-            "kw": keywords,
-            "v": self._compute_embedding_vector(keywords)
-        }
+        query_params = {"kw": keywords, "v": self._compute_embedding_vector(keywords)}
 
         logging.info('searching: ' + keywords + ', ' + 'approximate=' + str(approximate) + ', ' + 'hybrid=' + str(hybrid))
         logging.info('using search query: ' + q)
@@ -340,35 +307,35 @@ class Registry():
 
         # field', 'id', 'name', 'payload', 'score', 'type
         if approximate or hybrid:
-            results = [ {"name": result['name'], "type": result['type'], "id": result['id'], "scope": result['scope'], "score": result['score']} for result in results]
+            results = [{"name": result['name'], "type": result['type'], "id": result['id'], "scope": result['scope'], "score": result['score']} for result in results]
         else:
-            results = [ {"name": result['name'], "type": result['type'], "id": result['id'], "scope": result['scope']} for result in results]
+            results = [{"name": result['name'], "type": result['type'], "id": result['id'], "scope": result['scope']} for result in results]
 
         # do paging
         print(len(results))
-        page_results = results[page * page_size : ( page + 1 ) * page_size]
+        page_results = results[page * page_size : (page + 1) * page_size]
         return page_results
 
     ###### embeddings
-    def _init_search_embeddings_model(self):    
-        
+    def _init_search_embeddings_model(self):
+
         embeddings_model = self.properties['embeddings_model']
         logging.info('Loading embeddings model: ' + embeddings_model)
         self.embeddings_model = SentenceTransformer(embeddings_model)
 
         sentence = ['sample']
         embedding = self.embeddings_model.encode(sentence)[0]
-    
+
         # override vector_dimensions
         vector_dimensions = embedding.shape[0]
         self.properties['vector_dimensions'] = vector_dimensions
 
     def _compute_embedding_vector(self, text):
-        
-        sentence = [ text ]
+
+        sentence = [text]
         embedding = self.embeddings_model.encode(sentence)[0]
         return embedding.astype(np.float32).tobytes()
-    
+
     ###### registry functions
     def register_record(self, name, type, scope, description="", properties={}, rebuild=False):
         record = {}
@@ -378,13 +345,13 @@ class Registry():
         record['description'] = description
 
         record['properties'] = properties
-        
+
         # default contents
         record['contents'] = {}
 
         ## create a record on the registry name space
         p = self._get_record_path(name, scope)
-   
+
         self.connection.json().set(self._get_data_namespace(), p, record)
 
         # rebuild now
@@ -422,7 +389,6 @@ class Registry():
                 content = contents[key]
                 self.register_record_json(content, recursive=recursive, rebuild=rebuild)
 
-
     def update_record(self, name, type, scope, description="", properties={}, rebuild=False):
         record = {}
         record['name'] = name
@@ -433,7 +399,6 @@ class Registry():
         record['properties'] = properties
 
         return self.update_record_json(record, rebuild=rebuild)
-
 
     def update_record_json(self, record, recursive=True, rebuild=False):
         name = None
@@ -460,7 +425,7 @@ class Registry():
         return p
 
     def _get_scope_path(self, scope):
-        if scope[len(scope)-1] == '/':
+        if scope[len(scope) - 1] == '/':
             scope = scope[:-1]
 
         # compute json path given prefix, scope, and name
@@ -470,18 +435,17 @@ class Registry():
 
     def get_record(self, name, scope):
         p = self._get_record_path(name, scope)
-        record =  self.connection.json().get(self._get_data_namespace(), Path(p))
+        record = self.connection.json().get(self._get_data_namespace(), Path(p))
         if len(record) == 0:
             return {}
         else:
             record = record[0]
         return self.__get_json_value(record)
-    
+
     def get_record_data(self, name, scope, key, single=True):
         p = self._get_record_path(name, scope)
-        value =  self.connection.json().get(self._get_data_namespace(), Path(p + '.' + key))
+        value = self.connection.json().get(self._get_data_namespace(), Path(p + '.' + key))
         return self.__get_json_value(value, single=single)
-
 
     def set_record_data(self, name, scope, key, value, rebuild=False):
         p = self._get_record_path(name, scope)
@@ -549,7 +513,7 @@ class Registry():
     def get_records(self):
         contents = self.get_contents()
         records = []
-        r = json_utils.json_query(contents,"$..contents",single=False)
+        r = json_utils.json_query(contents, "$..contents", single=False)
         for ri in r:
             rs = list(ri.values())
             for rsi in rs:
@@ -564,7 +528,7 @@ class Registry():
         name = record['name']
         scope = record['scope']
 
-        # get full record so we can recursively delete 
+        # get full record so we can recursively delete
         record = self.get_record(name, scope)
         type = record['type']
 
@@ -577,9 +541,9 @@ class Registry():
 
     def list_records(self, type=None, scope="/"):
         sp = self._get_scope_path(scope)
-       
-        records =  self.connection.json().get(self._get_data_namespace(), Path(sp))[0]
-       
+
+        records = self.connection.json().get(self._get_data_namespace(), Path(sp))[0]
+
         if type:
             filtered_records = {}
             for key in records:
@@ -587,15 +551,15 @@ class Registry():
                 if record['type'] == type:
                     filtered_records[key] = record
             records = filtered_records
-    
+
         return records
 
     ######
     def _start(self):
         # logging.info('Starting session {name}'.format(name=self.name))
         self._start_connection()
-        
-         # initialize registry data
+
+        # initialize registry data
         self._init_registry_namespace()
 
         # build search index on registry
@@ -607,29 +571,30 @@ class Registry():
     def dumps(self, output_string):
         records = self.get_records()
         return str(records)
-    
+
     def dump(self, output_file):
         records = self.get_records()
-        with open(output_file, 'w') as fp:
-            json.dump(records, fp)
+        if os.path.exists(output_file):
+            with open(output_file, 'w') as fp:
+                json.dump(records, fp)
 
     def load(self, input_file):
-        with open(input_file, 'r') as fp:
-            records = json.load(fp)
+        if os.path.exists(input_file):
+            with open(input_file, 'r') as fp:
+                records = json.load(fp)
 
-            self._load_records(records)
+                self._load_records(records)
 
     def loads(self, input_string):
         records = json.loads(input_string)
 
         self._load_records(records)
-       
 
     def _load_records(self, records):
         print(records)
         for record in records:
             self.register_record_json(record)
-    
+
         # index registry
         self.build_index()
 
@@ -656,9 +621,9 @@ if __name__ == "__main__":
     parser.add_argument('--list', type=bool, default=False, action=argparse.BooleanOptionalAction, help='list records in the registry')
     parser.add_argument('--approximate', type=bool, default=False, action=argparse.BooleanOptionalAction, help='use approximate (embeddings) search')
     parser.add_argument('--hybrid', type=bool, default=False, action=argparse.BooleanOptionalAction, help='use hybrid (keyword and approximate) search')
- 
+
     args = parser.parse_args()
-   
+
     # set logging
     logging.getLogger().setLevel(args.loglevel.upper())
 
@@ -694,7 +659,7 @@ if __name__ == "__main__":
         print(records)
         for record in records:
             registry.update_record_json(record)
-    
+
         # index registry
         registry.build_index()
 
@@ -714,18 +679,11 @@ if __name__ == "__main__":
         results = registry.list_records()
         logging.info(results)
 
-
     #### SEARCH
     if args.search:
         keywords = args.search
 
         # search the registry
         results = registry.search_records(keywords, type=args.type, scope=args.scope, approximate=args.approximate, hybrid=args.hybrid, page=args.page, page_size=args.page_size)
-    
+
         logging.info(json.dumps(results, indent=4))
-
-   
-   
-
-  
-
